Enemy: Made level into a double

+++ HatQuest/HatQuest/Enemy.cs
@@ -35,13 +35,13 @@ namespace HatQuest
         /// <param name="width">The width of the enemy</param>
         /// <param name="height">The height of the enemy</param>
         /// <param name="player">The player</param>
-        public Enemy(EnemyType enemyType, int level, Point position, int width, int height, Player player) : base(enemyType.Sprite, position, width, height)
+        public Enemy(EnemyType enemyType, double level, Point position, int width, int height, Player player) : base(enemyType.Sprite, position, width, height)
         {
             name = enemyType.Name;
             this.player = player;
-            atk = enemyType.Attack * level;
-            def = enemyType.Defense * level;
-            maxHealth = currentHealth = enemyType.Health * level;
+            atk = (int)(enemyType.Attack * level);
+            def = (int)(enemyType.Defense * level);
+            maxHealth = currentHealth = (int)(enemyType.Health * level);
             abilities = enemyType.Abilities;
             random = EnemiesDirectory.random;
         }
